requirements.txt
Django==2.0.4
djangorestframework==3.8.2
Pillow==5.1.0
astroid==1.6.1
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pylint==1.8.2
pytz==2018.4
six==1.11.0
wrapt==1.10.11